web/slasher/app/requirements.txt
flask
Werkzeug
python-dotenv
bcrypt
uwsgi
